Fix breaking changes after rebase
