+++ experimental/PropertyDDS/examples/property-inspector/package.json
@@ -45,25 +45,25 @@
     "@fluid-experimental/property-proxy": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
     "@fluid-experimental/schemas": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
     "@fluidframework/aqueduct": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluid-internal/tree": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
     "@fluidframework/common-utils": "^0.32.1",
-    "@fluidframework/container-definitions": "^1.2.0",
-    "@fluidframework/container-loader": "^1.2.0",
-    "@fluidframework/container-runtime": "^1.2.0",
-    "@fluidframework/core-interfaces": "^1.2.0",
-    "@fluidframework/data-object-base": "^1.2.0",
-    "@fluidframework/datastore": "^1.2.0",
-    "@fluidframework/datastore-definitions": "^1.2.0",
-    "@fluidframework/map": "^1.2.0",
-    "@fluidframework/merge-tree": "^1.2.0",
-    "@fluidframework/protocol-definitions": "^0.1028.2000",
-    "@fluidframework/request-handler": "^1.2.0",
-    "@fluidframework/routerlicious-driver": "^1.2.0",
-    "@fluidframework/runtime-definitions": "^1.2.0",
-    "@fluidframework/sequence": "^1.2.0",
-    "@fluidframework/shared-object-base": "^1.2.0",
-    "@fluidframework/tinylicious-driver": "^1.2.0",
-    "@fluidframework/view-interfaces": "^1.2.0",
-    "@fluid-internal/tree": "1.2.0",
+    "@fluidframework/container-definitions": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/container-loader": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/container-runtime": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/core-interfaces": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/data-object-base": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/datastore": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/datastore-definitions": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/map": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/merge-tree": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/protocol-definitions": "^0.1029.1000-0",
+    "@fluidframework/request-handler": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/routerlicious-driver": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/runtime-definitions": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/sequence": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/shared-object-base": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/tinylicious-driver": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
+    "@fluidframework/view-interfaces": ">=2.0.0-internal.1.0.0 <2.0.0-internal.2.0.0",
     "@hig/fonts": "^1.0.2",
     "@material-ui/core": "^4.0.0",
     "@material-ui/lab": "^4.0.0-alpha.16",
