+++ experimental/PropertyDDS/packages/property-inspector-table/src/PropertyViews/Boolean.tsx
@@ -8,7 +8,7 @@ import Switch, { SwitchProps } from "@material-ui/core/Switch";
 import * as React from "react";
 import { IEditableValueCellProps } from "../EditableValueCell";
 import { IInspectorRow } from "../InspectorTableTypes";
-import { getPropertyValue } from "../utils";
+import { getPropertyValue } from "../propertyInspectorUtils";
 
 type BooleanProps = (IEditableValueCellProps & {
   onSubmit: (val: boolean, props: IEditableValueCellProps) => void;
